Moving sample handlers to example/ directory
